Add files via upload
You are given an array with n integers. You are to apply bucket sort to sort this array. Given a pre-defined number of equally-sized buckets b that you need to use, your task is to output the contents of each bucket at the end of the process.

You are expected to divide the input range into b equal-sized buckets. It is guaranteed that this is possible, for this problem. For example if the input array is [15,13,12,10,8,8,4] and b=4, then the bucket ranges will be [4,6],[7,9],[10,12] and [13,15]
Input
The first line contains two space-separated integers n (1≤n≤105) — the size of the array, and b (1≤b≤105) — the number of equal-sized buckets you are to use. The second line contains n integers a1,a2,…,an (−1012≤ai≤1012) — the elements of the array.

Output
Print b lines — each one corresponding to the contents of the b buckets after sorting, respectively. Each of these b lines should contain space-separated integers, the first of which would be k — the number of integers placed in that bucket. k integers follow in the same line — the contents of the bucket, in sorted order.